requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.11"
aiohttp==3.12.14 ; python_version >= "3.11"
aiosignal==1.4.0 ; python_version >= "3.11"
asyncio==3.4.3 ; python_version >= "3.11"
attrs==25.3.0 ; python_version >= "3.11"
frozenlist==1.7.0 ; python_version >= "3.11"
greenlet==3.2.3 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
idna==3.10 ; python_version >= "3.11"
multidict==6.6.3 ; python_version >= "3.11"
numpy==2.3.2 ; python_version >= "3.11"
pandas==2.3.1 ; python_version >= "3.11"
pathlib==1.0.1 ; python_version >= "3.11"
propcache==0.3.2 ; python_version >= "3.11"
psycopg2-binary==2.9.10 ; python_version >= "3.11"
pyarrow==21.0.0 ; python_version >= "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.11"
pytz==2025.2 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
sqlalchemy==2.0.41 ; python_version >= "3.11"
typing-extensions==4.14.1 ; python_version >= "3.11"
tzdata==2025.2 ; python_version >= "3.11"
yarl==1.20.1 ; python_version >= "3.11"
